requirements.txt
tensorflow==2.13
numpy
pillow
streamlit
scipy